separate project into lib and example
