+++ examples/3d_simple_character.rs
@@ -1,0 +1,3 @@
+use examples_common::*;
+
+fn main() {}
