+++ examples_common/src/camera/fly_camera.rs
