+++ examples_common/src/camera/orbit_camera.rs
