+++ examples_common/src/input.rs
