+++ examples_common/src/level/common.rs
