+++ examples_common/src/level/tracks/angled_walls.rs
