+++ examples_common/src/level/tracks/capsule_forest.rs
