+++ examples_common/src/level/tracks/cylinder_bridge.rs
