+++ examples_common/src/level/tracks/ground.rs
