+++ examples_common/src/level/tracks/half_height_obstacles.rs
