+++ examples_common/src/level/tracks/moving_platforms.rs
