+++ examples_common/src/level/tracks/narrow_beams.rs
