+++ examples_common/src/level/tracks/ramps.rs
