+++ examples_common/src/level/tracks/ridges.rs
