+++ examples_common/src/level/tracks/shape_obstacles.rs
