+++ examples_common/src/level/tracks/stairs.rs
