+++ examples_common/src/level/tracks/uneven_patches.rs
