+++ examples_common/src/lib.rs
@@ -1,3 +1,8 @@
+pub mod camera;
+pub mod input;
+pub mod level;
+pub mod movement;
+
 use avian3d::{PhysicsPlugins, prelude::*};
 use bevy::{
     pbr::{Atmosphere, light_consts::lux},
@@ -5,16 +10,30 @@ use bevy::{
     render::camera::Exposure,
 };
 use bevy_enhanced_input::prelude::Actions;
-use kcc_prototype::{
-    Attachments,
-    camera::FollowOffset,
-    camera::{CameraPlugin, MainCamera},
-    character::*,
-    input::{DefaultContext, InputPlugin},
-    input::{FlyCameraContext, OrbitCameraContext},
-    level::LevelGeneratorPlugin,
-    movement::{Character, KCCPlugin},
-};
+use camera::{CameraPlugin, FollowOffset, MainCamera};
+use input::{DefaultContext, FlyCameraContext, InputPlugin, OrbitCameraContext};
+use level::LevelGeneratorPlugin;
+use movement::{Character, KCCPlugin};
+
+pub const EXAMPLE_CHARACTER_RADIUS: f32 = 0.35;
+pub const EXAMPLE_CHARACTER_CAPSULE_LENGTH: f32 = 1.0;
+pub const EXAMPLE_MOVEMENT_SPEED: f32 = 8.0;
+pub const EXAMPLE_GROUND_ACCELERATION: f32 = 100.0;
+pub const EXAMPLE_AIR_ACCELERATION: f32 = 40.0;
+pub const EXAMPLE_FRICTION: f32 = 60.0;
+pub const EXAMPLE_WALKABLE_ANGLE: f32 = std::f32::consts::PI / 4.0;
+pub const EXAMPLE_JUMP_IMPULSE: f32 = 6.0;
+pub const EXAMPLE_GRAVITY: f32 = 20.0; // realistic earth gravity tend to feel wrong for games
+pub const EXAMPLE_STEP_HEIGHT: f32 = 0.25;
+pub const EXAMPLE_GROUND_CHECK_DISTANCE: f32 = 0.1;
+
+#[derive(Component)]
+#[relationship(relationship_target = Attachments)]
+pub struct AttachedTo(pub Entity);
+
+#[derive(Component)]
+#[relationship_target(relationship = AttachedTo)]
+pub struct Attachments(Vec<Entity>); // not sure about generaling this 
 
 fn main() -> AppExit {
     App::new()
